+++ R/RstoxFramework-package.R
@@ -15,21 +15,23 @@
 # Global variables
 utils::globalVariables(c(
 	 "..EDSUInfoToKeep", "..PSU", "..activeProcessID", "..areNumeric", "..digits",
-	 "..functionInputs", "..functionName", "..functionParameters", "..haulInfoToKeep", "..idCol",
-	 "..newProcessName", "..presentVariables", "..processDirty", "..propertyDirty",
-	 "..signifDigits", "..skipNAAt", "..stationInfoToKeep", "..subsetByNAOn_New",
-	 "..subsetByNAOn_Old", "..toKeep", "..validInd", "..x1x2y1y2", "AcoCat", "BeamKey",
-	 "BootstrapID", "BootstrapNetCDF4Data", "BootstrapSampleFactor", "CruiseKey", "DateTime",
-	 "Haul", "Layer", "LogKey", "LogOrigin", "LogOrigin2", "N", "PSU", "PolygonKey", "ProcessName",
-	 "ResampleFunction", "SampleUnit", "Station", "StationWeight", "StoX", "Stratum", "TableName",
-	 "WeightingFactor", "assignmentPasted", "canShowInMap", "capture.output", "ch",
-	 "currentVersion", "enabled", "freq", "functionInputProcessIDs", "functionInputs",
-	 "functionInputError", "functionName", "functionOutputDataType", "hasBeenRun", "hasProcessData",
-	 "lenght", "modelName", "name", "numStations", "oldCurrentVersion", "offset", "packageName",
-	 "possibleValues", "processDirty", "processID", "processIndex", "processName", "projectPath",
-	 "resampledCountWithUniqueName", "st_linestring", "st_sf", "st_sfc", "start_time", "tableName",
+	 "..functionInputs", "..functionInputs_UseProcessData", "..functionName",
+	 "..functionParameters", "..haulInfoToKeep", "..idCol", "..newProcessName",
+	 "..presentVariables", "..processDirty", "..propertyDirty", "..signifDigits", "..skipNAAt",
+	 "..stationInfoToKeep", "..subsetByNAOn_New", "..subsetByNAOn_Old",
+	 "..temporaryScaleFromResampling", "..validInd", "..x1x2y1y2", "AcoCat", "BeamKey",
+	 "BootstrapID", "BootstrapSampleFactor", "CruiseKey", "DateTime", "Haul", "Layer", "LogKey",
+	 "LogOrigin", "LogOrigin2", "N", "PSU", "PolygonKey", "ProcessName", "ResampleFunction",
+	 "SampleUnit", "Station", "StationWeight", "StoX", "Stratum", "WeightingFactor",
+	 "assignmentPasted", "canShowInMap", "ch", "currentVersion", "enabled", "freq",
+	 "functionInputError", "functionInputProcessIDs", "functionInputsRecursive",
+	 "functionInputs_UseProcessData", "functionName", "functionOutputDataType",
+	 "functionParameters", "hasBeenRun", "hasProcessData", "modelName", "name", "numStations",
+	 "offset", "oldCurrentVersion", "packageName", "possibleValues", "processDirty", "processID",
+	 "processIndex", "processName", "resamplingFactor", "start_time", "tableName",
 	 "terminalProcess", "transceiver", "usedInProcessIDs", "usedInProcessIndices",
-	 "usedInProcessNames", "value", "variableName", "verbose", "weightsPasted"))
+	 "usedInRecursiveProcessIndices", "usedInRecursiveProcessNames", "value", "variableName",
+	 "verbose", "weightsPasted"))
 
 .onLoad <- function(libname, pkgname) {
 	# Initiate the RstoxFramework environment: